Backend/requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.9.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.9
filelock==3.16.1
flake8==7.1.1
Flask==3.0.3
Flask-Cors==5.0.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
identify==2.6.2
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mccabe==0.7.0
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
virtualenv==20.27.1
Werkzeug==3.1.3
